refactor(circle): extract InCircleHeader, CircleInfoCard, JoinRequestsBanner (Sem 5 Día 3)
